refactor out autoconfig.js
